requirements.txt
ultralytics
gradio
opencv-python-headless
pillow
numpy
